+++ docs/_static/img/muon_logo.png
