+++ docs/_static/img/muon_logo_coloured.png
