+++ docs/index.md
@@ -7,20 +7,23 @@
 ```{toctree}
 :hidden: true
 :maxdepth: 1
-
-install.md
+:caption: Getting started
 tutorials.md
+notebooks/quickstart_mudata
+```
+
+```{toctree}
+:hidden: true
+:maxdepth: 1
+:caption: Documentation
+install.md
 io/input.md
 io/mudata.md
-io/output.md
 omics/uni.md
-omics/atac.md
-omics/citeseq.md
 omics/multi.md
+io/output.md
 api.md
 changelog.md
 contributing.md
 references.md
-
-notebooks/quickstart_mudata
 ```
